feat: Refactor GUI to use new HTTP classes and register observer
This changeset refactors the GUI of the application to use `MyHttpParameter` classes instead of `HttpParameter`. The `ContextMenu`, `ParserTableCellRenderer`, and `ParserTableModel` classes have been moved to new locations in the codebase, and the `parametersPanel.updateParserModel(parserModel)` call has been removed. Instead, the `parametersPanel` and `parserPanel` now share changes to the `requestWithMetadatas` list. Finally, this changeset registers the `MainTabGroup` class as an observer of the `ParserTableModel`, so that the visibility of the Parameters tab can be dynamically updated based on whether there is data to display.