Add bug report and feature request templates to the project
This change adds two new issue templates to the project, one for bug reports and another for feature requests. The bug report template includes fields for describing the bug and providing steps to reproduce it, as well as a section for providing environment information and error messages. The feature request template includes prompts for describing the problem or use case, the solution being requested, and any alternative solutions that have been considered.